feat: FAT32 driver mostly ready, not tested

+++ cpu/pages.c
@@ -3,12 +3,12 @@
 
 extern unsigned int _START;
 
-const u32 start_address = &_START;
+const uint32_t start_address = (uint32_t)&_START;
 
-u32 page_directory[1024] __attribute__((section(".multiboot.tables")))  __attribute__((aligned(4096))) = { 0 };
-u32 first_page_table[1024] __attribute__((section(".multiboot.tables"))) __attribute__((aligned(4096))) = { 0 };
-u32 second_page_table[1024] __attribute__((section(".multiboot.tables"))) __attribute__((aligned(4096))) = { 0 };
-u32 third_page_table[1024] __attribute__((section(".multiboot.tables"))) __attribute__((aligned(4096))) = { 0 };
+uint32_t page_directory[1024] __attribute__((section(".multiboot.tables")))  __attribute__((aligned(4096))) = { 0 };
+uint32_t first_page_table[1024] __attribute__((section(".multiboot.tables"))) __attribute__((aligned(4096))) = { 0 };
+uint32_t second_page_table[1024] __attribute__((section(".multiboot.tables"))) __attribute__((aligned(4096))) = { 0 };
+uint32_t third_page_table[1024] __attribute__((section(".multiboot.tables"))) __attribute__((aligned(4096))) = { 0 };
 
 void __attribute__((section(".multiboot.text"))) setup_dir(){
     for (int i = 0; i < 1024; ++i) {
@@ -25,8 +25,8 @@ void __attribute__((section(".multiboot.text"))) setup_table(){
         third_page_table[i] = (start_address + i * 0x001000 + 0x400000) | 3;
     }
 
-    page_directory[0]   = ((u32)first_page_table)  | 3;
-    page_directory[768] = ((u32)second_page_table) | 3;
-    page_directory[769] = ((u32)third_page_table)  | 3;
+    page_directory[0]   = ((uint32_t)first_page_table)  | 3;
+    page_directory[768] = ((uint32_t)second_page_table) | 3;
+    page_directory[769] = ((uint32_t)third_page_table)  | 3;
 
 }
